Script for generating SQL to renaming columns
Contains simple and more in-depth examples of scripts to help rename columns. Too often do we create a table with an import tool that gives the file headers to the columns. And too often do those contain characters forcing us to use brackets or quotes for our field names. This is a script I designed leveraging the system stored proc for renaming columns to help me in my everyday work.